+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [@adobe/spacecat-shared-google-client-v1.5.5](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.5.4...@adobe/spacecat-shared-google-client-v1.5.5) (2026-03-07)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1410](https://github.com/adobe/spacecat-shared/issues/1410)) ([e5b5634](https://github.com/adobe/spacecat-shared/commit/e5b5634f788a8f440ef49218c0b2fa1ccb42fb8a))
+
+## [@adobe/spacecat-shared-google-client-v1.5.4](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.5.3...@adobe/spacecat-shared-google-client-v1.5.4) (2026-03-02)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1223](https://github.com/adobe/spacecat-shared/issues/1223)) ([7ee8461](https://github.com/adobe/spacecat-shared/commit/7ee8461c99223d07a2f47bd6838b6942fcb30f28))
+
 ## [@adobe/spacecat-shared-google-client-v1.5.3](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.5.2...@adobe/spacecat-shared-google-client-v1.5.3) (2026-02-18)
 
 ### Bug Fixes